refactor: renamed the loading spinner to kebab case
